move GuessDial componentent into components
